+++ plugins/Variables/src/parse_miranda.cpp
@@ -486,7 +486,7 @@ static TCHAR *parseSpecialContact(ARGUMENTSINFO *ai)
 			return NULL;
 	}
 
-	int size = strlen(szProto) + _tcslen(szUniqueID) + 4;
+	size_t size = strlen(szProto) + _tcslen(szUniqueID) + 4;
 	TCHAR *res = (TCHAR*)mir_alloc(size * sizeof(TCHAR));
 	if (res == NULL) {
 		mir_free(szUniqueID);
